Use cached foms if present in directory
The cached file is `foms` in the directory containing the images.

For now the user has to remove the file in order to recalculate the
foms. Later on we may want to provide the command line option.

+++ src/foms.py
@@ -60,6 +60,11 @@ def write(outfile, cli_args, *args, **kwds):
         print(*args, **kwds, file=sys.stdout)
 
 def write_command(filename, images, commandStart, sphere_diameters,cli_args):
+    if Path(filename).is_file():
+        with open(filename, 'r') as infile:
+            for line in infile:
+                print (line.rstrip())
+        return
     with open(filename, 'w') as outfile:
         write(outfile,cli_args, '  ', end='')
         ds_header = ''.join(f'{d:7.1f}' for d in sphere_diameters)
